requirements_fixed.txt
# ===========================================
# DEPENDENCIAS PYTHON REFACTORIZADAS Y COMPATIBLES
# ===========================================
# Versiones compatibles y sin conflictos
# Fecha: 2025-01-17

# Core dependencies - Versiones compatibles
torch>=2.2.0,<2.3.0
transformers>=4.30.0,<4.40.0
accelerate>=0.20.0,<0.30.0
bitsandbytes>=0.39.0,<0.42.0

# NumPy - Versión estable compatible
numpy>=1.24.0,<2.0.0

# Llama specific - Versiones estables
llama-cpp-python>=0.2.0,<0.3.0
sentencepiece>=0.1.99,<0.2.0
protobuf>=3.20.0,<4.0.0

# Quantization support
auto-gptq>=0.4.0,<0.5.0
optimum>=1.12.0,<1.15.0

# Web server
fastapi>=0.100.0,<0.110.0
uvicorn>=0.22.0,<0.25.0
websockets>=11.0.0,<12.0.0

# HTTP client - Versión compatible
httpx>=0.24.1,<0.28.0

# Utilities - Versiones estables
requests>=2.31.0,<2.32.0
tqdm>=4.65.0,<4.67.0
psutil>=5.9.0,<6.0.0

# Machine Learning - Versiones compatibles
scikit-learn>=1.3.0,<1.4.0
pandas>=2.0.0,<2.2.0

# Embeddings y vectorización
sentence-transformers>=2.2.0,<2.3.0
faiss-cpu>=1.7.0,<1.8.0

# Base de datos
duckdb>=0.8.0,<0.9.0
asyncpg>=0.28.0,<0.29.0
psycopg2-binary>=2.9.0,<2.10.0

# Seguridad y encriptación
cryptography>=41.0.0,<42.0.0
pyotp>=2.8.0,<2.9.0
qrcode>=7.4.0,<7.5.0

# JWT y autenticación
PyJWT>=2.8.0,<2.9.0
bcrypt>=4.0.0,<4.1.0

# Monitoreo y logging
prometheus-client>=0.17.0,<0.18.0
structlog>=23.1.0,<23.2.0

# Desarrollo y testing
pytest>=7.4.0,<7.5.0
pytest-asyncio>=0.21.0,<0.22.0
black>=23.7.0,<23.8.0
flake8>=6.0.0,<6.1.0

# Visualización
matplotlib>=3.7.0,<3.8.0
seaborn>=0.12.0,<0.13.0

# Dependencias faltantes identificadas
numba>=0.57.0,<0.58.0

# Docker y deployment
docker>=6.1.0,<6.2.0
docker-compose>=1.29.0,<1.30.0

# JSON Schema - Versión compatible
jsonschema>=2.5.1,<3.0.0

# YAML - Versión compatible
PyYAML>=3.10,<5.0.0

# WebSocket client - Versión compatible
websocket-client>=0.32.0,<1.0.0

# Python dotenv - Versión compatible
python-dotenv>=0.13.0,<1.0.0

# Gradio - Versión compatible
gradio>=3.50.0,<4.0.0

# ChromaDB - Versión compatible
chromadb>=0.4.0,<1.1.0

# UMAP - Versión compatible
umap-learn>=0.5.0,<0.6.0

# PyNNDescent - Versión compatible
pynndescent>=0.5.0,<0.6.0
